[feat] 수익 결과를 담당하는 클래스 구현
WinningResult에서 ProfitResult로 클래스명 변경
